Added pull up refactoring
