+++ src/main/java/com/github/liuanxin/api/model/Document.java
@@ -1,0 +1,7 @@
+package com.github.liuanxin.api.model;
+
+public class Document {
+
+    protected String name;
+
+}
